+++ src/app/app.component.ts
@@ -3,8 +3,7 @@ import {PrimeNGConfig} from 'primeng/api';
 
 @Component({
     selector: 'app-root',
-    templateUrl: './app.component.html',
-    styleUrls: ['./app.component.scss']
+    templateUrl: './app.component.html'
 })
 export class AppComponent implements OnInit {
     menuMode = 'static';