+++ src/app/app.module.ts
@@ -110,7 +110,7 @@ import {TreeDemoComponent} from './demo/view/treedemo.component';
 import {PanelsDemoComponent} from './demo/view/panelsdemo.component';
 import {OverlaysDemoComponent} from './demo/view/overlaysdemo.component';
 import {MediaDemoComponent} from './demo/view/mediademo.component';
-import {MenusDemoComponent} from './demo/view/menusdemo.component';
+import {MenusComponent} from './demo/view/menus/menus.component';
 import {MessagesDemoComponent} from './demo/view/messagesdemo.component';
 import {MiscDemoComponent} from './demo/view/miscdemo.component';
 import {EmptyDemoComponent} from './demo/view/emptydemo.component';
@@ -138,6 +138,7 @@ import {PhotoService} from './demo/service/photoservice';
 import {ProductService} from './demo/service/productservice';
 import {BreadcrumbService} from './app.breadcrumb.service';
 import {MenuService} from './app.menu.service';
+import {ConfigService} from './demo/service/app.config.service';
 
 import dayGridPlugin from '@fullcalendar/daygrid';
 import timeGridPlugin from '@fullcalendar/timegrid';
@@ -260,7 +261,7 @@ FullCalendarModule.registerPlugins([
         PanelsDemoComponent,
         OverlaysDemoComponent,
         MediaDemoComponent,
-        MenusDemoComponent,
+        MenusComponent,
         MessagesDemoComponent,
         MessagesDemoComponent,
         MiscDemoComponent,
@@ -284,7 +285,7 @@ FullCalendarModule.registerPlugins([
     providers: [
         {provide: LocationStrategy, useClass: HashLocationStrategy},
         CountryService, CustomerService, EventService, IconService, NodeService,
-        PhotoService, ProductService, MenuService, BreadcrumbService
+        PhotoService, ProductService, MenuService, BreadcrumbService, ConfigService
     ],
     bootstrap: [AppComponent]
 })