+++ src/app/demo/view/listdemo.component.ts
@@ -6,7 +6,7 @@ import {BreadcrumbService} from '../../app.breadcrumb.service';
 
 @Component({
     templateUrl: './listdemo.component.html',
-    styleUrls: ['./listdemo.scss']
+    styleUrls: ['../../../assets/demo/badges.scss']
 })
 export class ListDemoComponent implements OnInit {