+++ src/app/demo/view/panelsdemo.component.ts
@@ -3,28 +3,7 @@ import {MenuItem} from 'primeng/api';
 import {BreadcrumbService} from '../../app.breadcrumb.service';
 
 @Component({
-    templateUrl: './panelsdemo.component.html',
-    styles: [`
-        :host ::ng-deep button {
-            margin-right: .25em;
-            margin-left: .25em;
-        }
-
-        :host ::ng-deep .p-splitbutton button {
-            margin-right: 0;
-            margin-left: 0;
-        }
-
-        :host ::ng-deep .p-splitter-panel-nested {
-            overflow: auto;
-        }
-
-        @media screen and (max-width: 960px) {
-            .card.toolbar-demo {
-                overflow: auto;
-            }
-        }
-    `]
+    templateUrl: './panelsdemo.component.html'
 })
 export class PanelsDemoComponent implements OnInit {