+++ src/app/pages/app.calendar.component.ts
@@ -1,6 +1,6 @@
 import {Component, OnInit} from '@angular/core';
 import {EventService} from '../demo/service/eventservice';
-
+import {BreadcrumbService} from '../app.breadcrumb.service';
 @Component({
     templateUrl: './app.calendar.component.html',
     styles: [`
@@ -8,6 +8,16 @@ import {EventService} from '../demo/service/eventservice';
             :host ::ng-deep .fc-header-toolbar {
                 display: flex;
                 flex-wrap: wrap;
+
+                .fc-dayGridMonth-button {
+                    margin-top: 1rem;
+                }
+                .fc-timeGridWeek-button{
+                    margin-top: 1rem;
+                }
+                .fc-timeGridDay-button{
+                    margin-top: 1rem;
+                }
             }
         }
     `]
@@ -26,7 +36,10 @@ export class AppCalendarComponent implements OnInit {
 
     clickedEvent = null;
 
-    constructor(private eventService: EventService) {
+    constructor(private eventService: EventService, public breadcrumbService: BreadcrumbService) {
+        this.breadcrumbService.setItems([
+            {label: 'Calendar'}
+        ]);
     }
 
     ngOnInit() {