+++ src/app/utilities/icons.component.ts
@@ -4,7 +4,6 @@ import {BreadcrumbService} from '../app.breadcrumb.service';
 
 @Component({
     templateUrl: './icons.component.html',
-    styleUrls: ['./icons.scss']
 })
 export class IconsComponent implements OnInit {